Remove underscores from .hbs files
It's only conventional to have an underscore for the _includes dir. We
should note that if you make any changes to css files, the page will
refresh. If we want hmr to work properly, we will need to extract all
css to a css dir. Right now, I believe having the templates and the
styles colocated is more beneficial than hmr.