Move eventListener binding into main.js
This is more clear than our previous init functions, because we can see
what is happening, as opposed to having to lookup what these init
functions do. It additionally prevents us from having to wrap all our
modules in funcs, which isn't really in the spirit of using modules. It
is all-around more readable.