Fix tooltip wrapping placement bug
The horizontal placement needs to be set before you set the vertical
placement. If you don't do this, when the tooltip is placed far left of
the screen and the tooltip wraps to avoid overflow-x, the newly
wrapped/expanded height won't be accessed in the top calculation. The
top calculation will reference the tooltip height under perfect
(unwrapped) conditions, and the tooltip will be placed awkwardly.

Since the left is being processed before everything else, it has been
removed from the state object. We could create additional state for
the left, but it isn't really necessary.